backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
imapclient==2.3.1
pyzmail36==1.0.4
openai==0.28.0
transformers==4.34.0
torch==2.0.1
nltk==3.8.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dateutil==2.8.2